requirements.txt
flask
flask_sse
pyjwt
python-dotenv
bcrypt
uuid
redis
gunicorn
bleach